refactor entity builder
