+++ FPS-ECS-Unity/Assets/_Project/Develop/Code/Runtime/Gameplay/ProgressionFeature/Systems/ProgressionInitializationSystem.cs
@@ -1,5 +1,6 @@
 using FpsEcs.Runtime.Gameplay.ProgressionFeature.Components;
 using FpsEcs.Runtime.Infrastructure.Factories;
+using FpsEcs.Runtime.Infrastructure.Factories.Entities;
 using Leopotam.EcsLite;
 using Leopotam.EcsLite.Di;
