+++ FPS-ECS-Unity/Assets/_Project/Develop/Code/Runtime/Infrastructure/Factories/Entities/IEntityBuilder.cs
@@ -1,0 +1,7 @@
+namespace FpsEcs.Runtime.Infrastructure.Factories.Entities
+{
+    public interface IEntityBuilder
+    {
+        IEntityBuilder With<T>() where T : struct;
+    }
+}
