+++ FPS-ECS-Unity/Assets/_Project/Develop/Code/Runtime/Infrastructure/Factories/Entities/IEntityBuilder.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 7f4a87893eb44baf9ce6b7f7cc1d5ffe
+timeCreated: 1759594495
