+++ FPS-ECS-Unity/Assets/_Project/Develop/Code/Runtime/Infrastructure/Factories/Entities/IEntityFactory.cs
@@ -1,0 +1,12 @@
+using Leopotam.EcsLite;
+using UnityEngine;
+
+namespace FpsEcs.Runtime.Infrastructure.Factories.Entities
+{
+    public interface IEntityFactory
+    {
+        void Initialize(EcsWorld world);
+        IEntityBuilder Create();
+        int CreateFrom(GameObject gameObject);
+    }
+}
